user-article delete

+++ 05_django_model_relation_l/articles/views.py
@@ -49,9 +49,11 @@ def detail(request, pk):
 
 @require_POST
 def delete(request, pk):
+    article = get_object_or_404(Article, pk=pk)
     if request.user.is_authticated:
-        article = get_object_or_404(Article, pk=pk)
-        article.delete()
+        if request.user == article.user:
+            article.delete()
+            return redirect('articles:index')
     return redirect('articles:index')
